[Compiles] Work on whole game module
Main files:
 (./src/game)
  - './game.gateway.ts' -> Server (handles events), gateway to clients
  - './game.service.ts' -> Provider, handles room and database (need to link with matchmaker)
 (./src/game/aliases)
  => every aliases used in the Game module
 (./src/game/dto)
  - './paddle.dto.ts' -> DTO information retrieved from client socket
 (./src/game/gameplay)
  - './gameplay.class.ts' -> actual gameplay manager, will do calculations
 (./src/game/room)
  - './gameroom.class.ts' -> Hosts the game, managing with room

The front will communicate via GameGateway. Its GameService instance will eventually assign a room to the entering socket (need a MatchMaking class).
Once two clients are in a GameRoom, its Gameplay instance will initialize a game.

For now this is just the strategy, need to implement it properly.